pyproject.toml
[project]
name = "chat-acp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "py-cord>=2.7.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.6",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
asyncio_default_fixture_loop_scope = "function"
